Definition of new data profile based on apn type
RIL_REQUEST_SETUP_DATA_CALL api has parameter profile id
which is optional as mentioned in the api descriptions in
ril.h. Though it is optional it is passed on always but
the values passed are inadequate for api types for MMS,
SUPL and HIPRI. These types will be used in RIL adaptation
to select a specific channel (an HSI channel for example)
for the SETUP request (a specific channel for MMS for example).
Differentiating it based on apn name is not always good as
there are operators (china mobile HK for example) providing
same apn for default and mms.

The value of the data profile is determined based on
the the apn type define in Phone.java. This patch add the
enumerations in framework base

Change-Id: If76bce9e5845a33e08221ac2f2ddfddd6d22618b
Signed-off-by: Alain Kouassu <alain.kouassu@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 32959